WIP to get OWM weather driver in a useable state.
The main issue with OWM is that they use internal ids to
identify locations and don't provide an easy way of searching
a "place" by various means (postal code, city, etc...).

This commit uses geonames.org to parse the location to get a lat/lon
pair, which is then used in the owm calls. Still fleshing out
some things...